backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
anthropic==0.42.0
yfinance==0.2.51
requests==2.32.3
